fix test

+++ webtau-junit5-examples/src/test/java/com/example/tests/junit5/DynamicTestsJavaTest.java
@@ -33,9 +33,9 @@ class DynamicTestsJavaTest {
     @TestFactory
     public Stream<DynamicTest> individualTestsUseGeneratedDisplayLabels() {
         return DynamicTests.fromTable(useCases, r -> {
-            long price = r.get("price");
-            long quantity = r.get("quantity");
-            long outcome = r.get("outcome");
+            int price = r.get("price");
+            int quantity = r.get("quantity");
+            int outcome = r.get("outcome");
 
             actual(PriceCalculator.calculate(price, quantity)).should(equal(outcome));
         });
